+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## vNEXT
-- Prevent voucher creation without value. (#32)
+- Mark assembly blocks as memory-safe and update deal price calculation syntax. (#33)
 - Harmonize SRLC wording. (#31)
 - Rename roles. (#29)
 - Make sponsored & non-sponsored amounts always divisible by deal volume to refund tasks fairly. (#28)